relocate ManifestUtils
