+++ ui/base-tool/src/main/java/org/cru/godtools/base/tool/databinding/TextViewAdapters.kt
@@ -12,7 +12,7 @@ import org.ccci.gto.android.common.picasso.widget.TextViewDrawableEndTarget
 import org.ccci.gto.android.common.picasso.widget.TextViewDrawableStartTarget
 import org.ccci.gto.android.common.util.dpToPixelSize
 import org.cru.godtools.base.tool.R
-import org.cru.godtools.base.tool.model.view.getTypeface
+import org.cru.godtools.base.tool.ui.util.getTypeface
 import org.cru.godtools.base.toolFileManager
 import org.cru.godtools.tool.model.Resource
 import org.cru.godtools.tool.model.Text
