+++ ui/base-tool/src/main/java/org/cru/godtools/base/tool/ui/util/ManifestUtils.kt
@@ -1,7 +1,7 @@
-package org.cru.godtools.base.tool.model.view
+package org.cru.godtools.base.tool.ui.util
 
 import android.content.Context
 import org.cru.godtools.base.ui.util.getTypeface
-import org.cru.godtools.xml.model.Manifest
+import org.cru.godtools.tool.model.Manifest
 
 fun Manifest.getTypeface(context: Context) = context.getTypeface(locale)
